feat: add Daily Operations store pulse visualizations
